Create Card server component
- as it does not have any client side functionality.
