+++ examples/react18-server-client-library/packages/my-lib/src/server/card/index.ts
@@ -1,0 +1,1 @@
+export * from "./card";
