+++ examples/react18-server-client-library/packages/my-lib/src/server/index.ts
@@ -1,0 +1,8 @@
+/**
+ * need to export server components and client components from separate files as
+ * directive on top of the file from which component is imported takes effect.
+ * i.e., server component re-exported from file with "use client" will behave as client component
+ * */
+
+// server component exports
+export * from "./card";
